requirements.txt
asgiref==3.7.2
astroid==3.0.2
black==23.12.1
cfgv==3.4.0
click==8.1.7
dill==0.3.7
distlib==0.3.8
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
filelock==3.13.1
identify==2.5.33
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pre-commit==3.6.0
psycopg2-binary==2.9.9
pylint==3.0.3
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
tomlkit==0.12.3
virtualenv==20.25.0
